# backend/internal/http/static/openapi.yaml
openapi: 3.0.3
info:
  title: Noryx API
  version: 1.0.0
  description: |
    The Noryx backend API, in full: every route the platform serves is
    described here, and a test fails the build when one is not.

    Entries generated from the router carry a summary derived from the
    handler's name; the ones that needed a human sentence have one. Both are
    checked against the router, so neither can drift.

    Authentication modes:
    - `Bearer` token (OIDC / Keycloak)
    - bootstrap compatibility header `X-Noryx-User` (lab mode)
servers:
  - url: /
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    xNoryxUser:
      type: apiKey
      in: header
      name: X-Noryx-User
  schemas:
...
    Extract:
      type: object
      properties:
        id:
          type: string
...
    ExtractMember:
      type: object
      properties:
        extractId:
          type: string
...
    ExtractMembersResponse:
      type: object
      properties:
        extract: { $ref: '#/components/schemas/Extract' }
        members:
          type: array
          items: { $ref: '#/components/schemas/ExtractMember' }
        subjects:
          type: array
          items: { type: string }
        shown: { type: integer }
        total: { type: integer }
...
    Ontology:
      type: object
      properties:
        id:
          type: string
    OntologyListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Ontology'
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Not enough permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /api/v1/ontologies/{ontologyID}/completeness:
    get:
      tags: [Datasets]
      summary: Get Ontology Completeness
      operationId: getOntologyCompleteness
      security:
        - bearerAuth: []
        - xNoryxUser: []
      parameters:
        - name: ontologyID
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OntologyListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/extracts/{extractID}:
    delete:
      tags: [Datasets]
      summary: Delete Extract
      operationId: deleteExtract
      security:
        - bearerAuth: []
        - xNoryxUser: []
      parameters:
        - name: extractID
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Extract removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/extracts/{extractID}/members:
    get:
      tags: [Datasets]
      summary: Get Extract Members
      operationId: getExtractMembers
      security:
        - bearerAuth: []
        - xNoryxUser: []
      parameters:
        - name: extractID
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The files the extract froze
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractMembersResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/ontologies/{ontologyID}/extracts:
    get:
      tags: [Datasets]
      summary: List Ontology Extracts
      operationId: listOntologyExtracts
      security:
        - bearerAuth: []
        - xNoryxUser: []
      parameters:
        - name: ontologyID
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OntologyListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Datasets]
      summary: Create Extract
      operationId: createExtract
      security:
        - bearerAuth: []
        - xNoryxUser: []
      parameters:
        - name: ontologyID
          in: path
          required: true
          schema: { type: string }
      responses:
        '201':
          description: Extract created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Extract'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...